refactor: usecases throwing exceptions

+++ src/app/core/use-cases/contact/DeleteContact.ts
@@ -1,3 +1,6 @@
+import { APIError } from "@/app/exceptions/base-error"
+import businessError from "@/app/exceptions/business-error"
+import { HttpStatusCode } from "@/app/exceptions/interfaces"
 import { ContactRepositoryDb } from "@/app/repository/ContactRepositoryDb"
 
 export class DeleteContact {
@@ -6,6 +9,14 @@ export class DeleteContact {
   }
 
   async execute (id: number) {
+    if (!id) {
+      throw new APIError("BAD_REQUEST",
+        HttpStatusCode.BAD_REQUEST,
+        true,
+        businessError.GENERIC,
+        undefined
+      )
+    }
     return this.contactRepository.delete(id)
   }
 }
