+++ src/app/core/use-cases/contract/DeleteContract.ts
@@ -1,3 +1,6 @@
+import { APIError } from "@/app/exceptions/base-error"
+import businessError from "@/app/exceptions/business-error"
+import { HttpStatusCode } from "@/app/exceptions/interfaces"
 import { ContractRepositoryDb, IContractRepositoryDbMethods } from "@/app/repository/ContractRepositoryDb"
 
 export class DeleteContract {
@@ -7,6 +10,14 @@ export class DeleteContract {
     }
 
     async execute (id: number, partnerId: number) {
+      if (!partnerId) {
+        throw new APIError("NOT_FOUND",
+          HttpStatusCode.NOT_FOUND,
+          true,
+          businessError.CONTRACT_NOT_FOUND,
+          undefined
+        )
+      }
       return this.contractRepository.deleteWithPartner(id, partnerId)
     }
 }
