+++ src/app/core/use-cases/partner/DeletePartner.ts
@@ -1,3 +1,6 @@
+import { APIError } from "@/app/exceptions/base-error"
+import businessError from "@/app/exceptions/business-error"
+import { HttpStatusCode } from "@/app/exceptions/interfaces"
 import { IPartnerRepositoryDbMethods, PartnerRepositoryDb } from "@/app/repository/PartnerRepositoryDb"
 
 export class DeletePartner {
@@ -7,6 +10,22 @@ export class DeletePartner {
     }
 
     async execute (id: number) {
-      return this.partnerRepository.delete(id)
+      if (!id) {
+        throw new APIError("NOT_FOUND",
+          HttpStatusCode.NOT_FOUND,
+          true,
+          businessError.MODALITY_NOT_FOUND,
+          undefined
+        )
+      }
+      const result = await this.partnerRepository.delete(id)
+      if (!result) {
+        throw new APIError("NOT_FOUND",
+          HttpStatusCode.NOT_FOUND,
+          true,
+          businessError.PARTNER_NOT_FOUND
+        )
+      }
+      return result
     }
 }
